hassan final

+++ NOOB/src/main/java/org/twspring/noob/Service/PcCentresService.java
@@ -20,20 +20,20 @@ public class PcCentresService {
     private final VendorRepository vendorRepository;
     private final AuthRepository authRepository;;
 
-
     public List<PcCentres> getAllPcCentres() {
         return pcCentresRepository.findAll();
     }
 
-
-    public void addPcCentres(PcCentres pcCentres,Integer vendorID) {
+    public void addPcCentres(PcCentres pcCentres, Integer vendorID) {
         Vendor vendor = vendorRepository.findVendorById(vendorID);
         if (vendor == null) {
             throw new ApiException("Vendor with ID " + vendorID + " not found");
         }
-
-
+        pcCentres.setApproved(false);
+        pcCentres.setVendor(vendor);
         pcCentresRepository.save(pcCentres);
+        //vendor.setPcCentres(Set.of(pcCentres2));
+        vendorRepository.save(vendor);
 
     }
 
@@ -42,13 +42,17 @@ public class PcCentresService {
         if (pcCentres1 == null) {
             throw new ApiException("PC Centre not found");
         }
-       pcCentres1.setCentreName(pcCentres.getCentreName());
+        if (pcCentres1.getVendor().getId() != pcCentres.getVendor().getId()) {
+            throw new ApiException("PC Centres don't match");
+        }
+        pcCentres1.setCentreName(pcCentres.getCentreName());
         pcCentres1.setLocation(pcCentres.getLocation());
         pcCentres1.setOpeningHours(pcCentres.getOpeningHours());
         pcCentres1.setContactNumber(pcCentres.getContactNumber());
         pcCentres1.setNumberOfPc(pcCentres.getNumberOfPc());
         pcCentres1.setRating(pcCentres.getRating());
-        pcCentres1.setApproved(false);
+        pcCentres1.setApproved(pcCentres.isApproved());
+
         pcCentresRepository.save(pcCentres1);
     }
 
@@ -57,50 +61,48 @@ public class PcCentresService {
         if (pcCentres == null) {
             throw new ApiException("PC Centres not found");
         }
+        if (pcCentres.getVendor().getId() != pcCentres.getVendor().getId()) {
+            throw new ApiException("PC Centres don't match");
+        }
         pcCentresRepository.delete(pcCentres);
     }
 
-    public List<PcCentres>getPcCentresByVendorID(Integer vendorId) {
-Vendor vendor=vendorRepository.findVendorById(vendorId);
-if (vendor==null){
-    throw new ApiException("Vendor with ID " + vendorId + " not found");
-}
-if (vendor.getPcCentres()==null){
-    throw new ApiException("PC Centres not found");
-}
+    public List<PcCentres> getPcCentresByVendorID(Integer vendorId) {
+        Vendor vendor = vendorRepository.findVendorById(vendorId);
+        if (vendor == null) {
+            throw new ApiException("Vendor with ID " + vendorId + " not found");
+        }
+        if (vendor.getPcCentres() == null) {
+            throw new ApiException("PC Centres not found");
+        }
+
 
         return pcCentresRepository.getPcCentresByVendorId(vendorId);
 
     }
 
-    //======
+    // ======
 
-    public void adminAprovedPcCenter(Integer adminId, Integer pcCenterId){
-        User u = authRepository.findUserById(adminId);
-        if (u==null){
-            throw new ApiException("User not found");
-        }else {
-            if (u.getRole().equalsIgnoreCase("ADMIN")) {
+    //@Transactional
+    public void adminAprovedPcCenter(Integer pcCenterId) {
+
                 PcCentres pcCentres = pcCentresRepository.findPcCentreById(pcCenterId);
                 if (pcCentres == null) {
                     throw new ApiException("PC Centre not found");
-                }else {
+                } else {
 
                     pcCentres.setApproved(true);
                     pcCentresRepository.save(pcCentres);
                 }
-            }else {
-                throw new ApiException("You are not an admin");
-            }
+
         }
-    }
 
-    ///=============
+    /// =============
 
-    public List<PcCentres>getPcCentresByRating(Integer rating){
-if (rating == null) {
-    throw new ApiException("Rating not found");
-}
+    public List<PcCentres> getPcCentresByRating(Integer rating) {
+        if (rating == null) {
+            throw new ApiException("Rating not found");
+        }
         return pcCentresRepository.findPcCentresByRating(rating);
     }
 
@@ -116,10 +118,21 @@ if (rating == null) {
         return pcCentres;
     }
 
-    public List<PcCentres>getPcCentreByLocation(String location) {
+    public List<PcCentres> getPcCentreByLocation(String location) {
         if (location == null) {
-        throw new ApiException("Location not found");
+            throw new ApiException("Location not found");
         }
         return pcCentresRepository.findPcCentresByLocation(location);
     }
+
+
+    public List<PcCentres>getApprovedPcCentres(){
+  return pcCentresRepository.findPcCentresByApprovedTrue();
+    }
+
+    public List<PcCentres>getAllNotApprovedPcCentre(){
+        return pcCentresRepository.findPcCentresByApprovedFalse();
+    }
+
+
 }
